fixed: fully working todays weather fragment

+++ app/src/main/java/com/example/weathertask/retrofit/TodaysWeatherJsonResponse.kt
@@ -9,10 +9,10 @@ class Clouds {
 
 class Coord {
     @SerializedName("lon")
-    var lon: Int? = null
+    var lon: Double? = null
 
     @SerializedName("lat")
-    var lat: Int? = null
+    var lat: Double? = null
 }
 
 class Main {
